+++ src/main/java/com/idark/valoria/client/ui/screen/HeavyWorkbenchScreen.java
@@ -164,8 +164,8 @@ public class HeavyWorkbenchScreen extends AbstractContainerScreen<HeavyWorkbench
             }
         }
 
-        craftableRecipes.sort(Comparator.comparing((b) -> b.recipe.getId()));
-        nonCraftableRecipes.sort(Comparator.comparing((b) -> b.recipe.getId()));
+        craftableRecipes.sort(Comparator.comparing((r -> r.recipe.getResultItem(RegistryAccess.EMPTY).getDescriptionId())));
+        nonCraftableRecipes.sort(Comparator.comparing((r -> r.recipe.getResultItem(RegistryAccess.EMPTY).getDescriptionId())));
         renderedRecipes.clear();
         renderedRecipes.addAll(craftableRecipes).addAll(nonCraftableRecipes);
     }