+++ src/main/java/com/idark/valoria/registries/entity/projectile/NatureArrow.java
@@ -5,6 +5,7 @@ import com.idark.valoria.client.particle.*;
 import com.idark.valoria.registries.*;
 import com.idark.valoria.util.*;
 import net.minecraft.resources.*;
+import net.minecraft.world.effect.*;
 import net.minecraft.world.entity.*;
 import net.minecraft.world.item.*;
 import net.minecraft.world.level.*;
@@ -53,6 +54,14 @@ public class NatureArrow extends AbstractProjectile implements TexturedArrow{
 
             ParticleEffects.smoothTrail(this.level(), target, pos, ColorParticleData.create(Pal.nature, Pal.seaGreen).build());
             ParticleEffects.trailMotionSparks(this.level(), pos, ColorParticleData.create(Pal.nature, Pal.seaGreen).build());
+        }
+    }
+
+    @Override
+    public void onHitEntity(EntityHitResult pResult){
+        super.onHitEntity(pResult);
+        if(pResult.getEntity() instanceof LivingEntity living){
+            living.addEffect(new MobEffectInstance(MobEffects.POISON, 10, 0));
         }
     }